+++ scripts/noir_drive.dart
@@ -12,7 +12,7 @@
 //
 //     capture [--ansi|--plain|--cells]   # --ansi paints the frame in color
 //     tree [depth]
-//     key <up|down|left|right|enter|tab|esc|backspace|pgup|pgdn|ctrl-<a-z>>
+//     key <up|down|left|right|enter|tab|space|esc|backspace|pgup|pgdn|ctrl-<a-z>>
 //     type <text...>
 //     click <x> <y>
 //     scroll <up|down|left|right> <x> <y>